# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portal 311",
        "description": "Documentación API Portal 311",
        "contact": {
            "name": "API Support",
            "url": "http://www.ogtic.gob.do",
            "email": "developer@ogtic.gob.do"
        },
        "version": "2.0"
    },
    "paths": {
        "/api/divisas": {
            "get": {
                "tags": [
                    "Divisas"
                ],
                "description": "Listado de divisas",
                "operationId": "divisa_index",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre de divisa a buscar",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "EUR"
                        }
                    },
                    {
                        "name": "ordercreated",
                        "in": "query",
                        "description": "Ordenar fecha de Creación de la divisa",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "ASC"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "Json",
                                            "example": "[...]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/divisa": {
            "post": {
                "tags": [
                    "Divisas"
                ],
                "description": "Crear una divisa nueva",
                "operationId": "divisa_store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre de la divisa",
                                        "type": "string",
                                        "example": "dolar"
                                    },
                                    "code": {
                                        "description": "Código de la divisa",
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "symbol": {
                                        "description": "Simbolo de la divisa",
                                        "type": "string",
                                        "example": "$"
                                    },
                                    "exchange_rate": {
                                        "description": "Tasa de cambio de la divisa",
                                        "type": "number",
                                        "format": "float",
                                        "example": "1.23456789"
                                    },
                                    "is_active": {
                                        "description": "Estado de la divisa",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "Json",
                                            "example": "[...]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/divisa/{id}": {
            "get": {
                "tags": [
                    "Divisas"
                ],
                "description": "Obtener una divisa por ID",
                "operationId": "divisa_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "Json",
                                            "example": "[...]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/divisa/{id}/update": {
            "post": {
                "tags": [
                    "Divisas"
                ],
                "description": "Actualizar una divisa",
                "operationId": "divisa_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre de la divisa",
                                        "type": "string",
                                        "example": "dolar"
                                    },
                                    "code": {
                                        "description": "Código de la divisa",
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "symbol": {
                                        "description": "Simbolo de la divisa",
                                        "type": "string",
                                        "example": "$"
                                    },
                                    "exchange_rate": {
                                        "description": "Tasa de cambio de la divisa",
                                        "type": "number",
                                        "format": "float",
                                        "example": "1.23456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "Json",
                                            "example": "[...]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/divisa/{id}/delete": {
            "delete": {
                "tags": [
                    "Divisas"
                ],
                "description": "Eliminar una divisa",
                "operationId": "divisa_destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Deleted"
                    },
                    "401": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/backoffice/v1/divisa/{divisa}/toggle": {
            "post": {
                "tags": [
                    "Divisas"
                ],
                "description": "Activar o desactivar una divisa",
                "operationId": "divisa_toggle",
                "parameters": [
                    {
                        "name": "divisa",
                        "in": "path",
                        "description": "divisa Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful Activated or Deactivated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "Json",
                                            "example": "[...]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error recurso no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Recurso no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "description": "Autenticacion",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string",
                                        "example": "adminp@pruebas.com"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "json",
                                            "example": "User information"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "bearer token for user"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Login"
                ],
                "description": "Salir del sistema",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "successful"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User has been logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            },
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "API-Key",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Divisas",
            "description": "Divisas"
        },
        {
            "name": "Login",
            "description": "Login"
        }
    ]
}